requirements.txt
torch==1.12.0+cu113 torchvision==0.13.0+cu113 torchaudio==0.12.0 --extra-index-url https://download.pytorch.org/whl/cu113
opencv-python
transformers==4.24.0
scikit-learn
pytorch_lightning==1.8.3.post1
pandas
timm

